# frontend/openapi/auth.json
{
  "components": {
    "schemas": {
      "Body_login_login_post": {
        "properties": {
          "client_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Id"
          },
          "client_secret": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Secret"
          },
          "grant_type": {
            "anyOf": [
              {
                "pattern": "password",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grant Type"
          },
          "password": {
            "title": "Password",
            "type": "string"
          },
          "scope": {
            "default": "",
            "title": "Scope",
            "type": "string"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "title": "Body_login_login_post",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "Message": {
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "title": "Message",
        "type": "object"
      },
...
      "ResetPassword": {
        "properties": {
          "new_password": {
            "title": "New Password",
            "type": "string"
          },
          "token": {
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "token",
          "new_password"
        ],
        "title": "ResetPassword",
        "type": "object"
      },
      "Token": {
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "token_type": {
            "title": "Token Type",
            "type": "string"
          }
        },
        "required": [
          "access_token",
          "token_type"
        ],
        "title": "Token",
        "type": "object"
      },
      "UserCreate": {
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          },
          "password": {
            "title": "Password",
            "type": "string"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "email",
          "password"
        ],
        "title": "UserCreate",
        "type": "object"
      },
      "UserPublic": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "disabled": {
            "default": false,
            "title": "Disabled",
            "type": "boolean"
          },
          "email": {
            "maxLength": 255,
            "minLength": 3,
            "title": "Email",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole",
            "default": "user"
          },
          "username": {
            "maxLength": 22,
            "minLength": 3,
            "title": "Username",
            "type": "string"
          },
          "verified": {
            "default": false,
            "title": "Verified",
            "type": "boolean"
          }
        },
        "required": [
          "username",
          "email",
          "id"
        ],
        "title": "UserPublic",
        "type": "object"
      },
      "UserRole": {
        "enum": [
          "user",
          "admin",
          "root"
        ],
        "title": "UserRole",
        "type": "string"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "description": "\n    A REST API service for user authentication.\n\n    This REST API:\n    - Allows users to register and login.\n    - Generates access tokens and refresh tokens.\n    - Allows users to refresh their access tokens.\n    - Allows users to logout.\n    - Allows users to view and revoke their refresh tokens.\n    ",
    "title": "Microservices [auth]",
    "version": "latest"
  },
  "openapi": "3.1.0",
  "paths": {
    "/login": {
      "post": {
        "description": "Login a user and return an access token.\n\nArgs:\n    response (Response): The response object.\n    request (Request): The request object.\n    ip_address (str): The client IP address.\n    session (AsyncSession): The database session.\n    form_data (OAuth2PasswordRequestForm): The form data.\n\nReturns:\n    Token: The access token.",
        "operationId": "login_login_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_login_login_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Token"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/password/forgot": {
      "post": {
        "description": "Send a password reset email.\n\nArgs:\n    session (AsyncSession): The database session.\n\nReturns:\n    Message: The response message.",
        "operationId": "send_forgot_password_password_forgot_post",
        "parameters": [
          {
            "in": "query",
            "name": "username_email",
            "required": true,
            "schema": {
              "title": "Username Email",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Send Forgot Password",
        "tags": [
          "auth"
        ]
      }
    },
    "/password/reset": {
      "post": {
        "description": "Reset user password.\n\nArgs:\n    session (AsyncSession): The database session.\n    token (str): The password reset token.\n    body (ResetPassword): Body containing the new password.\n\nReturns:\n    Message: The response message.",
        "operationId": "reset_password_password_reset_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPassword"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reset Password",
        "tags": [
          "auth"
        ]
      }
    },
    "/register": {
      "post": {
        "description": "Register a new user.\n\nArgs:\n    request (Request): The request object.\n    session (AsyncSession): The database session.\n    user (UserCreate): The user data.\n\nReturns:\n    UserPublic: The user data.",
        "operationId": "register_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register",
        "tags": [
          "auth"
        ]
      }
    },
...
  }
}